continue splitting
