+++ src/esp_light_mqtt.cpp
@@ -1,3 +1,18 @@
+/*
+  esp_light_mqtt.cpp - RGBW LED CONTROLLER
+  Copyright (C) 2020  Errol Sancaktar
+  This program is free software: you can redistribute it and/or modify
+  it under the terms of the GNU General Public License as published by
+  the Free Software Foundation, either version 3 of the License, or
+  (at your option) any later version.
+  This program is distributed in the hope that it will be useful,
+  but WITHOUT ANY WARRANTY; without even the implied warranty of
+  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+  GNU General Public License for more details.
+  You should have received a copy of the GNU General Public License
+  along with this program.  If not, see <http://www.gnu.org/licenses/>.
+*/
+
 #ifndef _esp_light_mqtt
 #define _esp_light_mqtt
 #include "config.h"
